gui/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pillow>=9.0.0
scipy>=1.9.0
numpy>=1.23.0
